requirements.txt
blessings==1.7
bpython==0.17.1
cachetools==2.1.0
cairocffi==1.0.2
CairoSVG==2.3.0
Cerberus==1.3.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cssselect2==0.2.1
curtsies==0.3.0
defusedxml==0.5.0
Django==2.0.6
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==2.2.0
django-dbbackup==3.2.0
django-extensions==2.0.7
django-filter==1.1.0
django-storages==1.7.2
django-taggit==0.22.2
djangorestframework==3.8.2
dropbox==9.4.0
factory-boy==2.11.1
Faker==0.8.15
ffmpeg-python==0.2.0
future==0.18.2
google-api-python-client==1.7.1
google-auth==1.5.0
google-auth-httplib2==0.0.3
greenlet==0.4.15
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
imgkit==1.0.1
Pillow==6.0.0
psutil==5.6.7
psycopg2-binary==2.8.3
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.19
Pygments==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.7.3
pytz==2018.4
requests==2.21.0
rsa==3.4.2
six==1.11.0
text-unidecode==1.2
tinycss2==1.0.2
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
WeasyPrint==46
webencodings==0.5.1
youtube-dl==2020.6.16.1
